Calculator Test

+++ src/main/java/com/drot/calculator/Calculator.java
@@ -4,6 +4,6 @@ public interface Calculator {
     int plus(int num1, int num2);
     int minus(int num1, int num2);
     int multiply(int num1, int num2);
-    float divide(float num1, float num2);
+    float divide(int num1, int num2);
 
 }
