+++ src/main/java/com/drot/calculator/CalculatorService.java
@@ -16,8 +16,9 @@ public class CalculatorService implements Calculator{
         return num1 * num2;
     }
 
-    public float divide(float num1, float num2) {
-            return num1 / num2;
+    public float divide(int num1, int num2) {
+            return (float) num1 / num2;
+
     }
